Rename cache-test folder to cache
