+++ client/extensions/wp-super-cache/state/reducer.js
@@ -6,12 +6,12 @@ import { combineReducers } from 'redux';
 /**
  * Internal dependencies
  */
-import cacheTest from './cache-test/reducer';
+import cache from './cache/reducer';
 import notices from './notices/reducer';
 import settings from './settings/reducer';
 
 export default combineReducers( {
-	cacheTest,
+	cache,
 	notices,
 	settings,
 } );
